requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=12.0.0

# Database
duckdb>=0.9.0

# Visualization
matplotlib>=3.6.0

# Configuration
pyyaml>=6.0

# Data download and processing
requests>=2.28.0
tqdm>=4.64.0

# Text processing
jellyfish>=1.0.0
rapidfuzz>=3.0.0
